BAU Move Notification related classes to dedicated package
