+++ app/uk/gov/hmrc/exports/services/notifications/NotificationService.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.exports.services
+package uk.gov.hmrc.exports.services.notifications
 
 import javax.inject.{Inject, Singleton}
 import play.api.Logger
