+++ app/uk/gov/hmrc/exports/services/notifications/WCOPointerMappingService.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.exports.services
+package uk.gov.hmrc.exports.services.notifications
 
 import com.github.tototoshi.csv.CSVReader
 import play.api.Logger
